fixed issues where gmail would hide/change content
When gmail detects two emails are very similar it hides all the content
that is shared between the two emails. If only some parts in an html
mail are shared these parts will be colored purple. The first thing is
fixed by adding invisble spans with the current time, and the second
should be fixed by changing the subject to include the current date